added: create and check avaialble endponints in booking-service, api-gateway properties edited for booking-service

+++ booking-service/src/main/java/com/hotel/booking/dto/response/BookingResponse.java
@@ -23,7 +23,7 @@ public class BookingResponse {
     private String roomType;
     private LocalDate checkInDate;
     private LocalDate checkOutDate;
-    private BigDecimal totalAmount;
+    private Float totalAmount;
     private BookingStatus status;
     private String guestName;
     private String guestEmail;
